replica, cmd/geth: Kafka event feeds
This commit is a step in the direction of running events for both
Websockets and Flume through Kafka.

On the producer side, it emits blocks and logs through the same
Kafka topic, with a terminator for each block to identify when
all the logs have been sent.

On the consumer side it aggregates the blocks and logs back into
ChainEvent objects, and handles chain re-orgs to emit any
removed logs events. The consumer side handles deduplication and
chain reorgs, on the assumption that it will be downstream from
multiple masters, which could end up in different positions of a
chain-reorg at a given moment. This could potentially lead to an
appearance of more chain reorgs, but will be eventually consistent
in terms of added and removed logs.

At the same time, I've cleaned up replica tests, which haven't been
touched in a while and were failling. I still need to add tests for
the kafka managed events.