cmd/geth: migratestate subcommand
This is the initial work on a migrate state subcommand.

Right now it will migrate the state from the last freezer block,
but that only works if the state was available for that particular
block (assuming that the state history goes back that far and that
specific block's state was committed).

My next update will be to pull all of the blocks from the last freezer
block forward to the latest block, then pull the state from the latest
block. Assuming we're working with a properly shutdown node, the latest
block staet should have been committed.