cmd/geth: Finish state pruner
The state pruner seems to be working now, based on Goerli. It turns
out that the state for the genesis block must be present in the database,
so we copy that now, as well as several other datapoints that geth looks
for to evaluate whether the database is current.

Additionally, we now copy over data not found in the freezer. Since this
is intended to become a part of the ethercattle snapshotting process,
we ideally want the state of the new database to match the state of the
old one, minus older state we're no longer concerned with. That way we
can resume from the same kafka offset and not have to worry about
missing blocks or data, or trying to figure out which offset we should
resume from (and ancient data tends to be older than we retain in Kafka
anyway).

With this in place, we will probably change to once a week we copy the
state from an old database to a new database, then update from kafka
after pruning, so that we will still have several hours of state data
in any servers generated from the snapshot.