replica: Kafka event producer reorg handling
When we built the event producer in the first place, it assumed that
every block would be emitted by the chain feed. It turns out that,
in the event of a chain reorg, the chain feed only emits the new
head, and not blocks that were never considered the latest block.

This means when the consumer tries to handle the reorg, it doesn't
know about some of the ancestor blocks, and thus can't handle it.

This change makes sure that the necessary information gets sent by
the producer so that the consumer has access to it. If the producer
gets a block for which it hasn't sent the parent, it will use the
database to work backwards to a known block and send any needed
blocks to the consumer.

Note that it won't send the same block twice - the producer feed
is not intended to fully reflect reorgs, just to provide the information
needed for the consumer to reconstruct the reorg. When a reorg
occurs, the consumer should have seen all of the blocks relating
to the reorg, but they shouldn't be expected to be in order for the
feed.