replica: Change offsetFeed from int64 to OffsetHash type
The offset feed previously only provided an offset, with no way to
map offsets to blocks. With a single producer one could naively assume
a 1:1 mapping of offsets to heads, but that doesn't always hold up
because of reorgs, and it totally falls apart with multiple producers
where you will have N times more offset events than blocks, where N
is the number of producers.

By including the hash with the offset, we can easily map between
emitted blocks and their corresponding offsets, even when it's not
a clear 1:1 mapping.