replica, internal/ethapi: Gas limits
Set the max gas limit for eth_call to 1000x the block gas limit,
and the default gas limit to 10x the block gas limit. That will allow
clients to specify that they need higher as limits, but we can use
our constraints engine to limit which customers are allowed to specify
high gas limits and on what contracts.