replica: Don't warm addresses if syncshutdown is True
The snapshotting process tends to take place on systems with less RAM,
so warming the address file into RAM can cause problems. In 1.9.12 we
didn't do this, because the warming process started later. Our
snapshotting logic requires the backend to be started during initialization,
but we don't really benefit from the warming process.