replica: Kafka events - fix log offset handling
For some reason I had been thinking that log.Index was the index within
the transaction, when it's actually the index within the block. So allocating
an array equal to the number of logs in a tx and then trying to insert
it by log.index fails beyond the most basic case. This adds a "logOffset"
attribute to receiptMeta, so we can index it by log.index - rmeta.logOffset