replica: Consumer may not process every block in the startup phase
During the startup phase of a KafkaEventConsumer, we roll back 5,000
offsets on each partition, with the intent that we prepopulate the CET
with any data it may need to process blocks that come in after the
starting offset. We already didn't emit these blocks, however we just
moved the real problem back 5,000 offsets, and could get stuck unable
to process blocks because the consumer started in the middle of something
5,000 offsets earlier.

This change sets cet.lastEmittedBlock back to empty after each record it
tries to emit prior to the starting offset. That will ensure that it will
continue prepopulating the CET, and not get hung up on blocks it doesn't
get see completely.