ethdb/cdc: Make required acks configurable
Previously the number of required acks was defaulting to 1, meaning
that the leader would respond as soon as it had recorded the message.
This means that if the leader is out of communication with the other
replicas (and perhaps is no longer the leader, but thinks it is) it
may respond that a message has been recorded that will removed when
that broker syncs up with the other replicas.

We're changing the default to require that the minimum number of in
sync replicas must have recorded the message, but adding a
`request.required.acks` parameter to the broker url to allow it to
be changed dynamically.