repica: Track skipped blocks rather than calling them finished
If we call blocks "finished" when they were actually skipped because
they were old, that can cause an outdated master coming online to
force the latest blocks to get garbage collected, beyond which we
cannot recover.

Instead, we use a skipped map, separate from the finished map. We
never clean up the skipped map, but each entry in the skipped map is
a 32 byte hash, so the entire history of the blockchain to date would
take up only a few hundred megs, and there aren't many scenarios where
we'd have more than a few thousand skipped blocks, so even without
cleanup this shouldn't be a practical issue.