replica: "finish" the siblings of the best pending child
When we process the pending child of a chain event, we were previously
ignoring any siblings of that child. The siblings would remain pending
indefinitely, which prevented reorg logic from being reached.

This only happened when:

* Multiple children of a single block were processed before their parent
* The getPendingChild() logic chose a block that later got reorged out

To handle this, we simply mark any siblings of an emitted block as finished
and no longer pending. Then if a child with a higher TD comes along,
the standard reorg logic handles it, rather than the pending child
logic (which assumes that there is still an ancestor that has not yet
been emitted).